requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
django-environ>=0.11.0
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
python-decouple>=3.8
django-filter>=23.3
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
tzdata>=2024.1